Module Owl_algodiff_graph_convert.Make

Parameters

module Core : Owl_algodiff_core_sig.Sig

Signature

val to_trace : Core.t list -> string

to_trace [t0; t1; ...] outputs the trace of computation graph on the terminal in a human-readable format.

val to_dot : Core.t list -> string

to_dot [t0; t1; ...] outputs the trace of computation graph in the dot file format which you can use other tools further visualisation, such as Graphviz.

val pp_num : Stdlib.Format.formatter -> Core.t -> unit

pp_num t pretty prints the abstract number used in Algodiff.