Module type Owl_computation_operator_sig.Sig

module Symbol : Owl_computation_symbol_sig.Sig

Vectorised functions

val noop : Symbol.Shape.Type.arr -> Symbol.Shape.Type.arr

TODO

val empty : int array -> Symbol.Shape.Type.arr

TODO

val zeros : int array -> Symbol.Shape.Type.arr

TODO

val ones : int array -> Symbol.Shape.Type.arr

TODO

val create : int array -> Symbol.Shape.Type.elt -> Symbol.Shape.Type.arr

TODO

val sequential : ?⁠a:Symbol.Shape.Type.elt -> ?⁠step:Symbol.Shape.Type.elt -> int array -> Symbol.Shape.Type.arr

TODO

val uniform : ?⁠a:Symbol.Shape.Type.elt -> ?⁠b:Symbol.Shape.Type.elt -> int array -> Symbol.Shape.Type.arr

TODO

val gaussian : ?⁠mu:Symbol.Shape.Type.elt -> ?⁠sigma:Symbol.Shape.Type.elt -> int array -> Symbol.Shape.Type.arr

TODO

val bernoulli : ?⁠p:Symbol.Shape.Type.elt -> int array -> Symbol.Shape.Type.arr

TODO

val init : int array -> (int -> Symbol.Shape.Type.elt) -> Symbol.Shape.Type.arr

TODO

val init_nd : int array -> (int array -> Symbol.Shape.Type.elt) -> Symbol.Shape.Type.arr

TODO

val shape : Symbol.Shape.Type.arr -> int array

TODO

val numel : Symbol.Shape.Type.arr -> int

TODO

val get : Symbol.Shape.Type.arr -> int array -> Symbol.Shape.Type.elt

TODO

val set : Symbol.Shape.Type.arr -> int array -> Symbol.Shape.Type.elt -> unit

TODO

val get_slice : int list list -> Symbol.Shape.Type.arr -> Symbol.Shape.Type.arr

TODO

val set_slice : int list list -> Symbol.Shape.Type.arr -> Symbol.Shape.Type.arr -> unit

TODO

val copy : Symbol.Shape.Type.arr -> Symbol.Shape.Type.arr

TODO

val copy_ : out:'a -> 'b -> 'c

TODO

val reset : Symbol.Shape.Type.arr -> unit

TODO

val reshape : Symbol.Shape.Type.arr -> int array -> Symbol.Shape.Type.arr

TODO

val reverse : Symbol.Shape.Type.arr -> Symbol.Shape.Type.arr

TODO

val tile : Symbol.Shape.Type.arr -> int array -> Symbol.Shape.Type.arr

TODO

val repeat : Symbol.Shape.Type.arr -> int array -> Symbol.Shape.Type.arr

TODO

val pad : ?⁠v:Symbol.Shape.Type.elt -> int list list -> Symbol.Shape.Type.arr -> Symbol.Shape.Type.arr

TODO

val expand : ?⁠hi:bool -> Symbol.Shape.Type.arr -> int -> Symbol.Shape.Type.arr

TODO

val squeeze : ?⁠axis:int array -> Symbol.Shape.Type.arr -> Symbol.Shape.Type.arr

TODO

val concatenate : ?⁠axis:int -> Symbol.Shape.Type.arr array -> Symbol.Shape.Type.arr

TODO

val concat : axis:int -> Symbol.Shape.Type.arr -> Symbol.Shape.Type.arr -> Symbol.Shape.Type.arr

TODO

val split : ?⁠axis:int -> 'a -> 'b -> 'c

TODO

val draw : ?⁠axis:int -> Symbol.Shape.Type.arr -> int -> Symbol.Shape.Type.arr * 'a array

TODO

val map : (Symbol.Shape.Type.elt -> Symbol.Shape.Type.elt) -> Symbol.Shape.Type.arr -> Symbol.Shape.Type.arr

TODO

val fold : ?⁠axis:int -> (Symbol.Shape.Type.elt -> Symbol.Shape.Type.elt -> Symbol.Shape.Type.elt) -> Symbol.Shape.Type.elt -> Symbol.Shape.Type.arr -> Symbol.Shape.Type.arr

TODO

val scan : ?⁠axis:int -> (Symbol.Shape.Type.elt -> Symbol.Shape.Type.elt -> Symbol.Shape.Type.elt) -> Symbol.Shape.Type.arr -> Symbol.Shape.Type.arr

TODO

val one_hot : int -> Symbol.Shape.Type.arr -> Symbol.Shape.Type.arr

TODO

val delay : (Symbol.Shape.Type.Device.A.arr -> Symbol.Shape.Type.Device.A.arr) -> Symbol.Shape.Type.arr -> Symbol.Shape.Type.arr

``delay f x`` returns ``f x``. It allows to use a function that is not tracked by the computation graph and delay its evaluation. The output must have the same shape as the input.

val delay_array : int array -> (Symbol.Shape.Type.Device.A.arr array -> Symbol.Shape.Type.Device.A.arr) -> Symbol.Shape.Type.arr array -> Symbol.Shape.Type.arr

``delay_array out_shape f x`` works in the same way as ``delay`` but is applied on an array of ndarrays. Needs the shape of the output as an argument.

val lazy_print : ?⁠max_row:int -> ?⁠max_col:int -> ?⁠header:bool -> ?⁠fmt:(Symbol.Shape.Type.Device.A.elt -> string) -> Symbol.Shape.Type.arr -> Symbol.Shape.Type.arr

``lazy_print x`` prints the output of ``x`` when it is evaluated. Is implemented as an identity node. For information about the optional parameters, refer to the ``print`` function of the ``Ndarray`` module.

val print : ?⁠max_row:'a -> ?⁠max_col:'b -> ?⁠header:'c -> ?⁠fmt:'d -> 'e -> unit

TODO

val abs : Symbol.Shape.Type.arr -> Symbol.Shape.Type.arr

TODO

val neg : Symbol.Shape.Type.arr -> Symbol.Shape.Type.arr

TODO

val floor : Symbol.Shape.Type.arr -> Symbol.Shape.Type.arr

TODO

val ceil : Symbol.Shape.Type.arr -> Symbol.Shape.Type.arr

TODO

val round : Symbol.Shape.Type.arr -> Symbol.Shape.Type.arr

TODO

val sqr : Symbol.Shape.Type.arr -> Symbol.Shape.Type.arr

TODO

val sqrt : Symbol.Shape.Type.arr -> Symbol.Shape.Type.arr

TODO

val log : Symbol.Shape.Type.arr -> Symbol.Shape.Type.arr

TODO

val log2 : Symbol.Shape.Type.arr -> Symbol.Shape.Type.arr

TODO

val log10 : Symbol.Shape.Type.arr -> Symbol.Shape.Type.arr

TODO

val exp : Symbol.Shape.Type.arr -> Symbol.Shape.Type.arr

TODO

val sin : Symbol.Shape.Type.arr -> Symbol.Shape.Type.arr

TODO

val cos : Symbol.Shape.Type.arr -> Symbol.Shape.Type.arr

TODO

val tan : Symbol.Shape.Type.arr -> Symbol.Shape.Type.arr

TODO

val sinh : Symbol.Shape.Type.arr -> Symbol.Shape.Type.arr

TODO

val cosh : Symbol.Shape.Type.arr -> Symbol.Shape.Type.arr

TODO

val tanh : Symbol.Shape.Type.arr -> Symbol.Shape.Type.arr

TODO

val asin : Symbol.Shape.Type.arr -> Symbol.Shape.Type.arr

TODO

val acos : Symbol.Shape.Type.arr -> Symbol.Shape.Type.arr

TODO

val atan : Symbol.Shape.Type.arr -> Symbol.Shape.Type.arr

TODO

val asinh : Symbol.Shape.Type.arr -> Symbol.Shape.Type.arr

TODO

val acosh : Symbol.Shape.Type.arr -> Symbol.Shape.Type.arr

TODO

val atanh : Symbol.Shape.Type.arr -> Symbol.Shape.Type.arr

TODO

val min : ?⁠axis:int -> Symbol.Shape.Type.arr -> Symbol.Shape.Type.arr

TODO

val max : ?⁠axis:int -> Symbol.Shape.Type.arr -> Symbol.Shape.Type.arr

TODO

val sum : ?⁠axis:int -> Symbol.Shape.Type.arr -> Symbol.Shape.Type.arr

TODO

val sum_reduce : ?⁠axis:int array -> Symbol.Shape.Type.arr -> Symbol.Shape.Type.arr

TODO

val signum : Symbol.Shape.Type.arr -> Symbol.Shape.Type.arr

TODO

val sigmoid : Symbol.Shape.Type.arr -> Symbol.Shape.Type.arr

TODO

val relu : Symbol.Shape.Type.arr -> Symbol.Shape.Type.arr

TODO

val min' : Symbol.Shape.Type.arr -> Symbol.Shape.Type.elt

TODO

val max' : Symbol.Shape.Type.arr -> Symbol.Shape.Type.elt

TODO

val sum' : Symbol.Shape.Type.arr -> Symbol.Shape.Type.elt

TODO

val l1norm' : Symbol.Shape.Type.arr -> Symbol.Shape.Type.elt

TODO

val l2norm' : Symbol.Shape.Type.arr -> Symbol.Shape.Type.elt

TODO

val l2norm_sqr' : Symbol.Shape.Type.arr -> Symbol.Shape.Type.elt

TODO

val clip_by_value : ?⁠amin:Symbol.Shape.Type.elt -> ?⁠amax:Symbol.Shape.Type.elt -> Symbol.Shape.Type.arr -> Symbol.Shape.Type.arr

TODO

val clip_by_l2norm : Symbol.Shape.Type.elt -> Symbol.Shape.Type.arr -> Symbol.Shape.Type.arr

TODO

val pow : Symbol.Shape.Type.arr -> Symbol.Shape.Type.arr -> Symbol.Shape.Type.arr

TODO

val scalar_pow : Symbol.Shape.Type.elt -> Symbol.Shape.Type.arr -> Symbol.Shape.Type.arr

TODO

val pow_scalar : Symbol.Shape.Type.arr -> Symbol.Shape.Type.elt -> Symbol.Shape.Type.arr

TODO

val atan2 : Symbol.Shape.Type.arr -> Symbol.Shape.Type.arr -> Symbol.Shape.Type.arr

TODO

val scalar_atan2 : Symbol.Shape.Type.elt -> Symbol.Shape.Type.arr -> Symbol.Shape.Type.arr

TODO

val atan2_scalar : Symbol.Shape.Type.arr -> Symbol.Shape.Type.elt -> Symbol.Shape.Type.arr

TODO

val hypot : Symbol.Shape.Type.arr -> Symbol.Shape.Type.arr -> Symbol.Shape.Type.arr

TODO

val min2 : Symbol.Shape.Type.arr -> Symbol.Shape.Type.arr -> Symbol.Shape.Type.arr

TODO

val max2 : Symbol.Shape.Type.arr -> Symbol.Shape.Type.arr -> Symbol.Shape.Type.arr

TODO

val add : Symbol.Shape.Type.arr -> Symbol.Shape.Type.arr -> Symbol.Shape.Type.arr

TODO

val sub : Symbol.Shape.Type.arr -> Symbol.Shape.Type.arr -> Symbol.Shape.Type.arr

TODO

val mul : Symbol.Shape.Type.arr -> Symbol.Shape.Type.arr -> Symbol.Shape.Type.arr

TODO

val div : Symbol.Shape.Type.arr -> Symbol.Shape.Type.arr -> Symbol.Shape.Type.arr

TODO

val add_scalar : Symbol.Shape.Type.arr -> Symbol.Shape.Type.elt -> Symbol.Shape.Type.arr

TODO

val sub_scalar : Symbol.Shape.Type.arr -> Symbol.Shape.Type.elt -> Symbol.Shape.Type.arr

TODO

val mul_scalar : Symbol.Shape.Type.arr -> Symbol.Shape.Type.elt -> Symbol.Shape.Type.arr

TODO

val div_scalar : Symbol.Shape.Type.arr -> Symbol.Shape.Type.elt -> Symbol.Shape.Type.arr

TODO

val scalar_add : Symbol.Shape.Type.elt -> Symbol.Shape.Type.arr -> Symbol.Shape.Type.arr

TODO

val scalar_sub : Symbol.Shape.Type.elt -> Symbol.Shape.Type.arr -> Symbol.Shape.Type.arr

TODO

val scalar_mul : Symbol.Shape.Type.elt -> Symbol.Shape.Type.arr -> Symbol.Shape.Type.arr

TODO

val scalar_div : Symbol.Shape.Type.elt -> Symbol.Shape.Type.arr -> Symbol.Shape.Type.arr

TODO

val fma : Symbol.Shape.Type.arr -> Symbol.Shape.Type.arr -> Symbol.Shape.Type.arr -> Symbol.Shape.Type.arr

TODO

val elt_equal : Symbol.Shape.Type.arr -> Symbol.Shape.Type.arr -> Symbol.Shape.Type.arr

TODO

val elt_not_equal : Symbol.Shape.Type.arr -> Symbol.Shape.Type.arr -> Symbol.Shape.Type.arr

TODO

val elt_less : Symbol.Shape.Type.arr -> Symbol.Shape.Type.arr -> Symbol.Shape.Type.arr

TODO

val elt_greater : Symbol.Shape.Type.arr -> Symbol.Shape.Type.arr -> Symbol.Shape.Type.arr

TODO

val elt_less_equal : Symbol.Shape.Type.arr -> Symbol.Shape.Type.arr -> Symbol.Shape.Type.arr

TODO

val elt_greater_equal : Symbol.Shape.Type.arr -> Symbol.Shape.Type.arr -> Symbol.Shape.Type.arr

TODO

val elt_equal_scalar : Symbol.Shape.Type.arr -> Symbol.Shape.Type.elt -> Symbol.Shape.Type.arr

TODO

val elt_not_equal_scalar : Symbol.Shape.Type.arr -> Symbol.Shape.Type.elt -> Symbol.Shape.Type.arr

TODO

val elt_less_scalar : Symbol.Shape.Type.arr -> Symbol.Shape.Type.elt -> Symbol.Shape.Type.arr

TODO

val elt_greater_scalar : Symbol.Shape.Type.arr -> Symbol.Shape.Type.elt -> Symbol.Shape.Type.arr

TODO

val elt_less_equal_scalar : Symbol.Shape.Type.arr -> Symbol.Shape.Type.elt -> Symbol.Shape.Type.arr

TODO

val elt_greater_equal_scalar : Symbol.Shape.Type.arr -> Symbol.Shape.Type.elt -> Symbol.Shape.Type.arr

TODO

val conv1d : ?⁠padding:Owl_types.padding -> Symbol.Shape.Type.arr -> Symbol.Shape.Type.arr -> int array -> Symbol.Shape.Type.arr

TODO

val conv2d : ?⁠padding:Owl_types.padding -> Symbol.Shape.Type.arr -> Symbol.Shape.Type.arr -> int array -> Symbol.Shape.Type.arr

TODO

val conv3d : ?⁠padding:Owl_types.padding -> Symbol.Shape.Type.arr -> Symbol.Shape.Type.arr -> int array -> Symbol.Shape.Type.arr

TODO

val transpose_conv1d : ?⁠padding:Owl_types.padding -> Symbol.Shape.Type.arr -> Symbol.Shape.Type.arr -> int array -> Symbol.Shape.Type.arr

TODO

val transpose_conv2d : ?⁠padding:Owl_types.padding -> Symbol.Shape.Type.arr -> Symbol.Shape.Type.arr -> int array -> Symbol.Shape.Type.arr

TODO

val transpose_conv3d : ?⁠padding:Owl_types.padding -> Symbol.Shape.Type.arr -> Symbol.Shape.Type.arr -> int array -> Symbol.Shape.Type.arr

TODO

val dilated_conv1d : ?⁠padding:Owl_types.padding -> Symbol.Shape.Type.arr -> Symbol.Shape.Type.arr -> int array -> int array -> Symbol.Shape.Type.arr

TODO

val dilated_conv2d : ?⁠padding:Owl_types.padding -> Symbol.Shape.Type.arr -> Symbol.Shape.Type.arr -> int array -> int array -> Symbol.Shape.Type.arr

TODO

val dilated_conv3d : ?⁠padding:Owl_types.padding -> Symbol.Shape.Type.arr -> Symbol.Shape.Type.arr -> int array -> int array -> Symbol.Shape.Type.arr

TODO

val max_pool1d : ?⁠padding:Owl_types.padding -> Symbol.Shape.Type.arr -> int array -> int array -> Symbol.Shape.Type.arr

TODO

val max_pool2d : ?⁠padding:Owl_types.padding -> Symbol.Shape.Type.arr -> int array -> int array -> Symbol.Shape.Type.arr

TODO

val max_pool3d : ?⁠padding:Owl_types.padding -> Symbol.Shape.Type.arr -> int array -> int array -> Symbol.Shape.Type.arr

TODO

val avg_pool1d : ?⁠padding:Owl_types.padding -> Symbol.Shape.Type.arr -> int array -> int array -> Symbol.Shape.Type.arr

TODO

val avg_pool2d : ?⁠padding:Owl_types.padding -> Symbol.Shape.Type.arr -> int array -> int array -> Symbol.Shape.Type.arr

TODO

val avg_pool3d : ?⁠padding:Owl_types.padding -> Symbol.Shape.Type.arr -> int array -> int array -> Symbol.Shape.Type.arr

TODO

val upsampling2d : Symbol.Shape.Type.arr -> int array -> Symbol.Shape.Type.arr

TODO

val conv1d_backward_input : Symbol.Shape.Type.arr -> Symbol.Shape.Type.arr -> int array -> Symbol.Shape.Type.arr -> Symbol.Shape.Type.arr

TODO

val conv1d_backward_kernel : Symbol.Shape.Type.arr -> Symbol.Shape.Type.arr -> int array -> Symbol.Shape.Type.arr -> Symbol.Shape.Type.arr

TODO

val conv2d_backward_input : Symbol.Shape.Type.arr -> Symbol.Shape.Type.arr -> int array -> Symbol.Shape.Type.arr -> Symbol.Shape.Type.arr

TODO

val conv2d_backward_kernel : Symbol.Shape.Type.arr -> Symbol.Shape.Type.arr -> int array -> Symbol.Shape.Type.arr -> Symbol.Shape.Type.arr

TODO

val conv3d_backward_input : Symbol.Shape.Type.arr -> Symbol.Shape.Type.arr -> int array -> Symbol.Shape.Type.arr -> Symbol.Shape.Type.arr

TODO

val conv3d_backward_kernel : Symbol.Shape.Type.arr -> Symbol.Shape.Type.arr -> int array -> Symbol.Shape.Type.arr -> Symbol.Shape.Type.arr

TODO

val transpose_conv1d_backward_input : Symbol.Shape.Type.arr -> Symbol.Shape.Type.arr -> int array -> Symbol.Shape.Type.arr -> Symbol.Shape.Type.arr

TODO

val transpose_conv1d_backward_kernel : Symbol.Shape.Type.arr -> Symbol.Shape.Type.arr -> int array -> Symbol.Shape.Type.arr -> Symbol.Shape.Type.arr

TODO

val transpose_conv2d_backward_input : Symbol.Shape.Type.arr -> Symbol.Shape.Type.arr -> int array -> Symbol.Shape.Type.arr -> Symbol.Shape.Type.arr

TODO

val transpose_conv2d_backward_kernel : Symbol.Shape.Type.arr -> Symbol.Shape.Type.arr -> int array -> Symbol.Shape.Type.arr -> Symbol.Shape.Type.arr

TODO

val transpose_conv3d_backward_input : Symbol.Shape.Type.arr -> Symbol.Shape.Type.arr -> int array -> Symbol.Shape.Type.arr -> Symbol.Shape.Type.arr

TODO

val transpose_conv3d_backward_kernel : Symbol.Shape.Type.arr -> Symbol.Shape.Type.arr -> int array -> Symbol.Shape.Type.arr -> Symbol.Shape.Type.arr

TODO

val dilated_conv1d_backward_input : Symbol.Shape.Type.arr -> Symbol.Shape.Type.arr -> int array -> int array -> Symbol.Shape.Type.arr -> Symbol.Shape.Type.arr

TODO

val dilated_conv1d_backward_kernel : Symbol.Shape.Type.arr -> Symbol.Shape.Type.arr -> int array -> int array -> Symbol.Shape.Type.arr -> Symbol.Shape.Type.arr

TODO

val dilated_conv2d_backward_input : Symbol.Shape.Type.arr -> Symbol.Shape.Type.arr -> int array -> int array -> Symbol.Shape.Type.arr -> Symbol.Shape.Type.arr

TODO

val dilated_conv2d_backward_kernel : Symbol.Shape.Type.arr -> Symbol.Shape.Type.arr -> int array -> int array -> Symbol.Shape.Type.arr -> Symbol.Shape.Type.arr

TODO

val dilated_conv3d_backward_input : Symbol.Shape.Type.arr -> Symbol.Shape.Type.arr -> int array -> int array -> Symbol.Shape.Type.arr -> Symbol.Shape.Type.arr

TODO

val dilated_conv3d_backward_kernel : Symbol.Shape.Type.arr -> Symbol.Shape.Type.arr -> int array -> int array -> Symbol.Shape.Type.arr -> Symbol.Shape.Type.arr

TODO

val max_pool1d_backward : Owl_types.padding -> Symbol.Shape.Type.arr -> int array -> int array -> Symbol.Shape.Type.arr -> Symbol.Shape.Type.arr

TODO

val max_pool2d_backward : Owl_types.padding -> Symbol.Shape.Type.arr -> int array -> int array -> Symbol.Shape.Type.arr -> Symbol.Shape.Type.arr

TODO

val max_pool3d_backward : Owl_types.padding -> Symbol.Shape.Type.arr -> int array -> int array -> Symbol.Shape.Type.arr -> Symbol.Shape.Type.arr

TODO

val avg_pool1d_backward : Owl_types.padding -> Symbol.Shape.Type.arr -> int array -> int array -> Symbol.Shape.Type.arr -> Symbol.Shape.Type.arr

TODO

val avg_pool2d_backward : Owl_types.padding -> Symbol.Shape.Type.arr -> int array -> int array -> Symbol.Shape.Type.arr -> Symbol.Shape.Type.arr

TODO

val avg_pool3d_backward : Owl_types.padding -> Symbol.Shape.Type.arr -> int array -> int array -> Symbol.Shape.Type.arr -> Symbol.Shape.Type.arr

TODO

val upsampling2d_backward : Symbol.Shape.Type.arr -> int array -> Symbol.Shape.Type.arr -> Symbol.Shape.Type.arr

TODO

val row_num : Symbol.Shape.Type.arr -> int

TODO

val col_num : Symbol.Shape.Type.arr -> int

TODO

val row : Symbol.Shape.Type.arr -> 'a -> Symbol.Shape.Type.arr

TODO

val rows : Symbol.Shape.Type.arr -> int array -> Symbol.Shape.Type.arr

TODO

val copy_row_to : Symbol.Shape.Type.arr -> 'a -> 'b -> unit

TODO

val copy_col_to : Symbol.Shape.Type.arr -> 'a -> 'b -> unit

TODO

val diag : ?⁠k:int -> Symbol.Shape.Type.arr -> Symbol.Shape.Type.arr

TODO

val trace : Symbol.Shape.Type.arr -> Symbol.Shape.Type.elt

TODO

val dot : Symbol.Shape.Type.arr -> Symbol.Shape.Type.arr -> Symbol.Shape.Type.arr

TODO

val transpose : ?⁠axis:int array -> Symbol.Shape.Type.arr -> Symbol.Shape.Type.arr

TODO

val to_rows : Symbol.Shape.Type.arr -> 'a array

TODO

val of_rows : Symbol.Shape.Type.arr array -> Symbol.Shape.Type.arr

TODO

val to_cols : Symbol.Shape.Type.arr -> 'a array

TODO

val of_cols : Symbol.Shape.Type.arr array -> Symbol.Shape.Type.arr

TODO

val of_array : Symbol.Shape.Type.elt array -> int array -> Symbol.Shape.Type.arr

TODO

val of_arrays : Symbol.Shape.Type.elt array array -> Symbol.Shape.Type.arr

TODO

val to_arrays : Symbol.Shape.Type.arr -> Symbol.Shape.Type.elt array array

TODO

Scalar functions

module Scalar : sig ... end

module Mat : sig ... end

module Linalg : sig ... end