Module Owl_const

Metric system: CGS, MKS, SI, and physical constants.

Values of physical constants CGS < MKS < SI. Read wikipedia on CGS and SI system for more details.

International System of Units (French: Système international d'unités, SI), historically also called the MKSA system of units for metre–kilogram–second–ampere.

The SI system of units extends the MKS system and has 7 base units, by expressing any measurement of physical quantities using fundamental units of Length, Mass, Time, Electric Current, Thermodynamic Temperature, Amount of substance and Luminous Intensity, which are Metre, Kilogram, Second, Ampere, Kelvin, Mole and Candela respectively.

http://www.npl.co.uk/upload/pdf/units-of-measurement-poster.pdf

Maths constants

val e : float

e = 2.718281828459045235360287471352662498

val euler : float

euler = 0.577215664901532860606512090082402431

val log2e : float

log2e = 1.442695040888963407359924681001892137

val log10e : float

log10e = 0.434294481903251827651128918916605082

val loge2 : float

loge2 = 0.693147180559945309417232121458176568

val loge10 : float

loge10 = 2.302585092994045684017991454684364208

val logepi : float

logepi = 1.144729885849400174143427351353058711

val sqrt1_2 : float

sqrt1_2 = 0.707106781186547524400844362104849039

val sqrt2 : float

sqrt2 = 1.414213562373095048801688724209698079

val sqrt3 : float

sqrt3 = 1.732050807568877293527446341505872366

val sqrtpi : float

sqrtpi = 1.772453850905516027298167483341145182

val pi : float

pi = 3.141592653589793238462643383279502884

val pi2 : float

pi2 = 6.283185307179586476925286766559005768

val pi4 : float

pi4 = 12.56637061435917295385057353311801153

val pi_2 : float

pi_2 = 1.570796326794896619231321691639751442

val pi_4 : float

pi_4 = 0.785398163397448309615660845819875721

val eps : float

eps = 1e-15

Constants depending on Bigarray kind

val zero : ('a, 'b) Stdlib.Bigarray.kind -> 'a

zero kind returns value zero of the given number type kind.

val one : ('a, 'b) Stdlib.Bigarray.kind -> 'a

one kind returns value one of the given number type kind.

val neg_one : ('a, 'b) Stdlib.Bigarray.kind -> 'a

neg_one kind returns negative one of the given number type kind.

val pos_inf : ('a, 'b) Stdlib.Bigarray.kind -> 'a

pos_inf kind returns positive infinity of the given number type kind.

val neg_inf : ('a, 'b) Stdlib.Bigarray.kind -> 'a

neg_inf kind returns negative infinity of the given number type kind.

val min_float32 : float

Miminum value of single precision float number, i.e. ~-.340282346638528859811704183484516925440.0

val max_float32 : float

Maximum value of single precision float number, i.e. 340282346638528859811704183484516925440.0

val min_float64 : float

Miminum value of double precision float number.

val max_float64 : float

Maximum value of double precision float number.

Unit prefixes

module Prefix : sig ... end

SI: International System of Units

module SI : sig ... end

MKS: MKS system of units

module MKS : sig ... end

CGS: Centimetre–gram–second system of units

module CGS : sig ... end

CGSM: Unit Systems in Electromagnetism

module CGSM : sig ... end