Module Neuron.AlphaDropout

type neuron_typ = {

mutable rate : float;

mutable in_shape : int array;

mutable out_shape : int array;

}

Neuron type definition.

val create : float -> neuron_typ

Create the neuron.

val connect : int array -> neuron_typ -> unit

Connect this neuron to others in a neural network.

val copy : neuron_typ -> neuron_typ

Make a deep copy of the neuron and its parameters.

val run : Optimise.Algodiff.t -> neuron_typ -> Optimise.Algodiff.t

Execute the computation in this neuron.

val to_string : neuron_typ -> string

Convert the neuron to its string representation. The string is often a summary of the parameters defined in the neuron.

val to_name : unit -> string

Return the name of the neuron.