Module Optimise.Utils

Utils module

val sample_num : Algodiff.t -> int

Return the total number of samples in passed in ndarray.

val draw_samples : Algodiff.t -> Algodiff.t -> int -> Algodiff.t * Algodiff.t

draw_samples x y draws samples from both x (observations) and y (labels). The samples will be drew along axis 0, so x and y must agree along axis 0.

val get_chunk : Algodiff.t -> Algodiff.t -> int -> int -> Algodiff.t * Algodiff.t

get_chunk x y i c gets a continuous chunk of c samples from position i from x (observations) and y (labels).