Parameter Make.1-Neuron

module Optimise : Owl_optimise_generic_sig.Sig

Init neuron

module Init : sig ... end

Input neuron

module Input : sig ... end

Activation neuron

module Activation : sig ... end

Linear neuron

module Linear : sig ... end

LinearNoBias neuron

module LinearNoBias : sig ... end

Recurrent neuron

module Recurrent : sig ... end

LSTM neuron

module LSTM : sig ... end

GRU neuron

module GRU : sig ... end

Conv1D neuron

module Conv1D : sig ... end

Conv2D neuron

module Conv2D : sig ... end

Conv3D neuron

module Conv3D : sig ... end

DilatedConv1D neuron

module DilatedConv1D : sig ... end

DilatedConv2D neuron

module DilatedConv2D : sig ... end

DilatedConv3D neuron

module DilatedConv3D : sig ... end

TransposeConv1D neuron

module TransposeConv1D : sig ... end

TransposeConv2D neuron

module TransposeConv2D : sig ... end

TransposeConv3D neuron

module TransposeConv3D : sig ... end

FullyConnected neuron

module FullyConnected : sig ... end

MaxPool1D neuron

module MaxPool1D : sig ... end

MaxPool2D neuron

module MaxPool2D : sig ... end

AvgPool1D neuron

module AvgPool1D : sig ... end

AvgPool2D neuron

module AvgPool2D : sig ... end

GlobalMaxPool1D neuron

module GlobalMaxPool1D : sig ... end

GlobalMaxPool2D neuron

module GlobalMaxPool2D : sig ... end

GlobalAvgPool1D neuron

module GlobalAvgPool1D : sig ... end

GlobalAvgPool2D neuron

module GlobalAvgPool2D : sig ... end

UpSampling1D neuron

module UpSampling1D : sig ... end

UpSampling2D neuron

module UpSampling2D : sig ... end

UpSampling3D neuron

module UpSampling3D : sig ... end

Padding1D neuron

module Padding1D : sig ... end

Padding2D neuron

module Padding2D : sig ... end

Padding3D neuron

module Padding3D : sig ... end

Lambda neuron

module Lambda : sig ... end

LambdaArray neuron

module LambdaArray : sig ... end

Dropout neuron

module Dropout : sig ... end

Reshape neuron

module Reshape : sig ... end

Flatten neuron

module Flatten : sig ... end

Add neuron

module Add : sig ... end

Mul neuron

module Mul : sig ... end

Dot neuron

module Dot : sig ... end

Max neuron

module Max : sig ... end

Average neuron

module Average : sig ... end

Concatenate neuron

module Concatenate : sig ... end

Normalisation neuron

module Normalisation : sig ... end

GaussianNoise neuron

module GaussianNoise : sig ... end

GaussianDropout neuron

module GaussianDropout : sig ... end

AlphaDropout neuron

module AlphaDropout : sig ... end

Embedding neuron

module Embedding : sig ... end

Masking neuron

module Masking : sig ... end

Core functions

type neuron =

| Input of Input.neuron_typ

| Linear of Linear.neuron_typ

| LinearNoBias of LinearNoBias.neuron_typ

| Embedding of Embedding.neuron_typ

| LSTM of LSTM.neuron_typ

| GRU of GRU.neuron_typ

| Recurrent of Recurrent.neuron_typ

| Conv1D of Conv1D.neuron_typ

| Conv2D of Conv2D.neuron_typ

| Conv3D of Conv3D.neuron_typ

| DilatedConv1D of DilatedConv1D.neuron_typ

| DilatedConv2D of DilatedConv2D.neuron_typ

| DilatedConv3D of DilatedConv3D.neuron_typ

| TransposeConv1D of TransposeConv1D.neuron_typ

| TransposeConv2D of TransposeConv2D.neuron_typ

| TransposeConv3D of TransposeConv3D.neuron_typ

| FullyConnected of FullyConnected.neuron_typ

| MaxPool1D of MaxPool1D.neuron_typ

| MaxPool2D of MaxPool2D.neuron_typ

| AvgPool1D of AvgPool1D.neuron_typ

| AvgPool2D of AvgPool2D.neuron_typ

| GlobalMaxPool1D of GlobalMaxPool1D.neuron_typ

| GlobalMaxPool2D of GlobalMaxPool2D.neuron_typ

| GlobalAvgPool1D of GlobalAvgPool1D.neuron_typ

| GlobalAvgPool2D of GlobalAvgPool2D.neuron_typ

| UpSampling2D of UpSampling2D.neuron_typ

| Padding2D of Padding2D.neuron_typ

| Dropout of Dropout.neuron_typ

| Reshape of Reshape.neuron_typ

| Flatten of Flatten.neuron_typ

| Lambda of Lambda.neuron_typ

| LambdaArray of LambdaArray.neuron_typ

| Activation of Activation.neuron_typ

| GaussianNoise of GaussianNoise.neuron_typ

| GaussianDropout of GaussianDropout.neuron_typ

| AlphaDropout of AlphaDropout.neuron_typ

| Normalisation of Normalisation.neuron_typ

| Add of Add.neuron_typ

| Mul of Mul.neuron_typ

| Dot of Dot.neuron_typ

| Max of Max.neuron_typ

| Average of Average.neuron_typ

| Concatenate of Concatenate.neuron_typ
Types of neuron.

val get_in_out_shape : neuron -> int array * int array

Get both input and output shapes of a neuron.

val get_in_shape : neuron -> int array

Get the input shape of a neuron.

val get_out_shape : neuron -> int array

Get the output shape of a neuron.

val connect : int array array -> neuron -> unit

Connect this neuron to others in a neural network.

val init : neuron -> unit

Initialise the neuron and its parameters.

val reset : neuron -> unit

Reset the parameters in a neuron.

val mktag : int -> neuron -> unit

Tag the neuron, used by ``Algodiff`` module.

val mkpar : neuron -> Optimise.Algodiff.t array

Assemble all the trainable parameters in an array, used by ``Optimise`` module.

val mkpri : neuron -> Optimise.Algodiff.t array

Assemble all the primal values in an array, used by ``Optimise`` module.

val mkadj : neuron -> Optimise.Algodiff.t array

Assemble all the adjacent values in an array, used by ``Optimise`` module.

val update : neuron -> Optimise.Algodiff.t array -> unit

Update trainable parameters in a neuron, used by ``Optimise`` module.

val load_weights : neuron -> Optimise.Algodiff.t array -> unit

Load both trainable and non-trainable parameters into the neuron.

val save_weights : neuron -> Optimise.Algodiff.t array

Assemble both trainable and non-trainable parameters of the neuron.

val copy : neuron -> neuron

Make a deep copy of the neuron and its parameters.

val run : Optimise.Algodiff.t array -> neuron -> Optimise.Algodiff.t

Execute the computation in this neuron.

val to_string : neuron -> string

Convert the neuron to its string representation. The string is often a summary of the parameters defined in the neuron.

val to_name : neuron -> string

Return the name of the neuron.