Module Neuron.Activation

type typ =

| Elu

| Relu

| Sigmoid

| HardSigmoid

| Softmax of int

| Softplus

| Softsign

| Tanh

| Relu6

| LeakyRelu of float

| TRelu of float

| Custom of Optimise.Algodiff.t -> Optimise.Algodiff.t

| None
Types of activation functions.

type neuron_typ = {

mutable activation : typ;

mutable in_shape : int array;

mutable out_shape : int array;

}

Neuron type definition.

val create : typ -> neuron_typ

Create the neuron.

val connect : int array -> neuron_typ -> unit

Connect this neuron to others in a neural network.

val run_activation : Optimise.Algodiff.t -> typ -> Optimise.Algodiff.t

Run one specific activation function.

val copy : neuron_typ -> neuron_typ

Make a deep copy of the neuron and its parameters.

val run : Optimise.Algodiff.t -> neuron_typ -> Optimise.Algodiff.t

Execute the computation in this neuron.

val activation_to_string : typ -> string

Return the name of a specific activation function.

val to_string : neuron_typ -> string

Convert the neuron to its string representation. The string is often a summary of the parameters defined in the neuron.

val to_name : unit -> string

Return the name of the neuron.