Module Sig.Normalisation

type neuron_typ = {

mutable axis : int;

mutable beta : Optimise.Algodiff.t;

mutable gamma : Optimise.Algodiff.t;

mutable mu : Optimise.Algodiff.t;

mutable var : Optimise.Algodiff.t;

mutable decay : Optimise.Algodiff.t;

mutable training : bool;

mutable in_shape : int array;

mutable out_shape : int array;

}

Neuron type definition.

val create : ?⁠training:bool -> ?⁠decay:float -> ?⁠mu:Optimise.Algodiff.A.arr -> ?⁠var:Optimise.Algodiff.A.arr -> int -> neuron_typ

Create the neuron. Note that axis 0 is the batch axis.

val connect : int array -> neuron_typ -> unit

Connect this neuron to others in a neural network.

val init : neuron_typ -> unit

Initialise the neuron and its parameters.

val reset : neuron_typ -> unit

Reset the parameters in a neuron.

val mktag : int -> neuron_typ -> unit

Tag the neuron, used by ``Algodiff`` module.

val mkpar : neuron_typ -> Optimise.Algodiff.t array

Assemble all the trainable parameters in an array, used by ``Optimise`` module.

val mkpri : neuron_typ -> Optimise.Algodiff.t array

Assemble all the primial values in an array, used by ``Optimise`` module.

val mkadj : neuron_typ -> Optimise.Algodiff.t array

Assemble all the adjacent values in an array, used by ``Optimise`` module.

val update : neuron_typ -> Optimise.Algodiff.t array -> unit

Update trainable parameters of the neuron, used by ``Optimise`` module.

val load_weights : neuron_typ -> Optimise.Algodiff.t array -> unit

Load both trainable and non-trainable parameters into the neuron.

val save_weights : neuron_typ -> Optimise.Algodiff.t array

Assemble both trainable and non-trainable parameters of the neuron.

val copy : neuron_typ -> neuron_typ

Make a deep copy of the neuron and its parameters.

val run : Optimise.Algodiff.t -> neuron_typ -> Optimise.Algodiff.t

Execute the computation in this neuron.

val to_string : neuron_typ -> string

Convert the neuron to its string representation. The string is often a summary of the parameters defined in the neuron.

val to_name : unit -> string

Return the name of the neuron.