Module Owl_types

This module defines the types shared by various sub-libraries in Owl. Note that they just wrappers, to find the exact module signature, please refer to the definition in the corresponding module.

include module type of struct include Owl_types_common end

type number = Owl_types_common.number =

| F32

| F64

| C32

| C64

type ('a, 'b) owl_arr = ('a, 'b, Stdlib.Bigarray.c_layout) Stdlib.Bigarray.Genarray.t

type index = Owl_types_common.index =

| I of int

| L of int list

| R of int list

type slice = index list

type index_ = Owl_types_common.index_ =

| I_ of int

| L_ of int array

| R_ of int array

type slice_ = index_ array

type padding = Owl_types_common.padding =

| SAME

| VALID

type device_type = Owl_types_common.device_type =

| CPU

| OpenCL

| CUDA

module type Ndarray_Basic = sig ... end

module type Ndarray_Compare = sig ... end

module type Ndarray_Mutable = sig ... end

module type Ndarray_Algodiff = sig ... end

module type Ndarray_Numdiff = sig ... end

module type Stats_Dist = sig ... end

module type Computation_Device = sig ... end