Module Owl_view

View module This module is used to create views atop of an ndarray. The view creation is very light-weighted and avoids copying actual data. You can further create views atop of existing views using slicing functions.

All the views share the same underlying ndarray and any modification will be reflected on the original ndarray.

module Make (A : Owl_types.Ndarray_Basic) : sig ... end