Module Owl_computation_engine

This functor takes a device as its input, then it generates the computation graph module without flattening the module hierarchy.

module Make_Graph : functor (Device : Owl_types_computation_device.Sig) -> sig ... end

module Flatten : functor (Engine : Owl_types_computation_engine.Sig) -> sig ... end