Module Owl_pretty

Pretty print the n-dimensional array

Ndarray pretty printer

val pp_dsnda : Stdlib.Format.formatter -> ('a, 'b, 'c) Stdlib.Bigarray.Genarray.t -> unit

``pp_dsnda`` is the pretty printer for n-dimensional arrays.

val dsnda_to_string : ?⁠header:bool -> ?⁠max_row:int -> ?⁠max_col:int -> ?⁠elt_to_str_fun:('a -> string) -> ('a, 'b, Stdlib.Bigarray.c_layout) Stdlib.Bigarray.Genarray.t -> string

``dsnda_to_string x`` converts ``x`` into a string for pretty printing

val print_dsnda : ?⁠header:bool -> ?⁠max_row:int -> ?⁠max_col:int -> ?⁠elt_to_str_fun:('a -> string) -> ?⁠formatter:Stdlib.Format.formatter -> ('a, 'b, Stdlib.Bigarray.c_layout) Stdlib.Bigarray.Genarray.t -> unit

``print_dsnda`` provides the better control of pretty printing function of owl's n-dimensional array. max_row and max_col specify the maximum number of rows and columns to display. header specifies whether or not to print out the headers. fmt is the function to format every element into string.

Dataframe pretty printer

val pp_dataframe : Stdlib.Format.formatter -> Owl_dataframe.t -> unit

``pp_dataframe`` is the pretty printer for dataframe.

val dataframe_to_string : ?⁠header:bool -> ?⁠names:string array -> ?⁠max_row:int -> ?⁠max_col:int -> ?⁠elt_to_str_fun:(Owl_dataframe.elt -> string) -> Owl_dataframe.t -> string

``dataframe_to_string x`` converts ``x`` into a string for pretty printing

val print_dataframe : ?⁠header:bool -> ?⁠names:string array -> ?⁠max_row:int -> ?⁠max_col:int -> ?⁠elt_to_str_fun:(Owl_dataframe.elt -> string) -> Stdlib.Format.formatter -> Owl_dataframe.t -> unit

``print_dataframe x`` converts ``x`` into a string for pretty printing