Module Owl_types_operator

Operator definitions such as add, sub, mul, and div. This signature defines the functions need to be implemented.

module type BasicSig = sig ... end

module type ExtendSig = sig ... end

module type MatrixSig = sig ... end

module type NdarraySig = sig ... end

module type LinalgSig = sig ... end